requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.2.2
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
